back/requirements.txt
asgiref==3.11.0
asttokens==3.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
decorator==5.2.1
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.1
idna==3.11
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
numpy==2.3.5
parso==0.8.5
prompt_toolkit==3.0.52
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.4
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.3
urllib3==2.6.2
wcwidth==0.2.14
